==> 03_Methyl_Differential.Rmd <==
---
title: "Methylation - Differential methylation & DMRs"
format: html
---

## Notes

### Meta analysis

-   Inverse variance weighting allows for correlated variables to be combined.

## Code

### Libraries

```{r}

# Plotting and graphics
library(ggplot2)
library(RColorBrewer)
library(scales)
library(colorspace)
library(patchwork)
library(ggrepel)
library(ggpubr)

# Data QC 
library(wateRmelon)
library(DNAmArray)
library(minfi)
library(maxprobes)

# Annotation
library(IlluminaHumanMethylationEPICanno.ilm10b4.hg19)
library(org.Hs.eg.db)
library(AnnotationDbi)
library(TxDb.Hsapiens.UCSC.hg19.knownGene)

# Programming, parsing and data wrangling
library(mgsub)
library(GenomicFeatures)
#library(doSNOW) # parallel processes on Windows
library(foreach)
library(RcppEigen) # For faster lm 
library(GenomicRanges)

# Statistics and analyses
library(limma)
library(factoextra)
library(sva)
library(pvca)
library(dmrff)
library(DMRcate)
library(splines)

# Gene ontology
library(clusterProfiler)
#library(methylGSA)

```

### Parameters

```{r}
#set the treat log2 treshold
trt_log2fc_thr <- 0.1
maxgap_val <- 1000

```

### Directories

```{r}

# Defined main dir
scripts_dir <- getwd()
main_dir <- paste0(scripts_dir, "/methylation/")
data_dir <- paste0(scripts_dir, "/rawdata/")
res_dir <- paste0(scripts_dir, "/results/")
pca_dir <- paste0(scripts_dir, "/pca/")
dmr_dir <- paste0(res_dir, "/dmr/")
de_dir <- paste0(scripts_dir, "/differential_methylation/")
gocpg_dir <- paste0(de_dir, "go/")
godmr_dir <- paste0(dmr_dir, "go/")
util_dir <- paste0(scripts_dir, "/utilities/")
log_dir <- paste0(scripts_dir, "/logs/")

# Ensure outputs are created
x <- suppressMessages(sapply(c(dmr_dir, gocpg_dir, godmr_dir), dir.create, recursive=TRUE, showWarnings=FALSE)); rm(x)

```

### Source R scripts

```{r}

# Source script that sources files
source("install/source_rscripts.R")

# Source relevant scripts
SourceExternalScripts(paste0(util_dir, "R/"), "*.R$", ignore.case=FALSE)

```

## Read in the preprocessed data

```{r}

preProccessedData <- readRDS(paste0(data_dir,"preproccessed_methyl.RDS"))
qnt.mSetSqFlt <- preProccessedData$qnt.mSetSqFlt
fun.mSetSqFlt <- preProccessedData$fun.mSetSqFlt
metadata <- preProccessedData$metadata

# Get M-normalised matrix
## Functional
fun.m_mat <- minfi::getM(fun.mSetSqFlt)
fun.m_mat <- as.matrix(na.omit(fun.m_mat))
fun.m_mat <- fun.m_mat[order(rowVars(fun.m_mat), decreasing=TRUE),]

# Load data
data(IlluminaHumanMethylationEPICanno.ilm10b4.hg19)
annotdata <- getAnnotation(IlluminaHumanMethylationEPICanno.ilm10b4.hg19)

```

#### Read in metadata

```{r}

# Read in metadata from EDA script
metadata <- read.table(paste0(res_dir,"methylation_sample_metadata.tsv"), sep="\t", header=TRUE)
rownames(metadata) <- metadata$Sample_Name

```
`
#### Prep design/metadata

```{r}

# Make factor names match
for_de.metadata <- metadata

# Set up design matrix
rownames(for_de.metadata) <- for_de.metadata$Sample_Name
bmicont.design <- model.matrix(~ 0 + Batch + sex + Z_MAT_BMI + Z_STAGE_WEEKS + Z_MAT_AGE + Z_RRNA_ABSORB +  Z_DNAul + Z_TISSUEmg , data=for_de.metadata)


```

#### Probe-wise

For Treat method and ref for below statements, see: https://www.ncbi.nlm.nih.gov/pmc/articles/PMC2654802/

Treat method tests for LFC \> a threshold, rather than 0.

Note that a low threshold needs to be chosen as the method requires that LFCs be significantly different from this, so the effective threshold would likely be greater than it.

Returned p-values are slightly conservative, due to it innately testing a composite NH. Therefore FDR is likely to be better than the rate returned by the p-values. However this means that the NH p-value distribution doesn't follow strict uniformity and is biased towards larger values. This means that while BH adjustment is still viable, some aren't (ie https://academic.oup.com/jrsssb/article/67/4/555/7110602).

```{r}

# Make contrast matrix

# Fit lm

bmicont.design <- model.matrix(~ 0 +  sex + Batch + Z_MAT_BMI + Z_STAGE_WEEKS + Z_MAT_AGE + Z_RRNA_ABSORB +  Z_DNAul + Z_TISSUEmg , data=for_de.metadata)
bmicont.fit <- lmFit(fun.m_mat[,rownames(bmicont.design)],bmicont.design)

#make contrast matrix
bmicont.contMatrix <- makeContrasts(Z_MAT_AGE,
                                    Z_MAT_BMI,
                                    Z_STAGE_WEEKS,
                                    sexmale-sexfemale,
                                    levels=bmicont.design)
rownames(bmicont.contMatrix) <- gsub("Intercept","(Intercept)",rownames(bmicont.contMatrix))

# fit the contrasts
bmicont.fit <- contrasts.fit(bmicont.fit, bmicont.contMatrix)

# Shrink variance
eby.bmicont.fit <- eBayes(bmicont.fit, trend=FALSE, robust=TRUE)

# Treat variance (ie lfc > 0)
trt.bmicont.fit <- treat(bmicont.fit, lfc=trt_log2fc_thr)

```

### Non linear stage-weeks using splines

```{r}
#allow df=5 as a reasonably sized time course
X <- ns(metadata$Z_STAGE_WEEKS, df=5)

#replace stage_weeks with the splines
spline.design <- model.matrix(~ 0 +  sex + Batch + Z_MAT_BMI + X + Z_MAT_AGE + Z_RRNA_ABSORB +  Z_DNAul + Z_TISSUEmg , data=for_de.metadata)


  # Fit the linear model
fit <- lmFit(fun.m_mat[,rownames(spline.design)], spline.design)

# Test for average time effect
contrast.matrix <- makeContrasts(Stage_spline = (X1 + X2 + X3 + X4 + X5)/5, levels = spline.design)

# Fit contrasts
fit2 <- contrasts.fit(fit, contrast.matrix)

# Apply empirical Bayes moderation
fit2 <- eBayes(fit2, trend=FALSE, robust=TRUE)

#And treat like before
fit2 <- treat(fit2, lfc=trt_log2fc_thr)

# Get the differentially expressed genes
splineDF <- GetLimmaResults(fit2,contrast.matrix, annotdata,trt_toggle = TRUE)

#bonferroni adjust
splineDF$bonferroni <- p.adjust(splineDF$P.Value,method="bonferroni")

spline_coefficients <- fit$coefficients[, grep("^X", colnames(fit$coefficients))]

#get the significant probes
splineDF_sig <- splineDF[ splineDF$bonferroni <= 0.05,]


topExamples <- splineDF_sig %>%
dplyr::slice_min(bonferroni,n=6) %>%
dplyr::pull(Name)

fun.b_mat <- as.matrix(minfi::getBeta(fun.mSetSqFlt))
dat <- fun.b_mat[topExamples, ]
# Create a long format data frame with CpG names included
long_df <- data.frame(stage_weeks = metadata$stage_weeks) %>%
cbind(t(dat)) %>%
tidyr::pivot_longer(cols = -stage_weeks, names_to = "CpG", values_to = "value")

g <- ggplot(long_df, aes(x = stage_weeks, y = value, color = CpG)) +
geom_point() +  # Points for each CpG
labs(x = "Developmental Stage (weeks)", y = "Methylation (beta)") + facet_wrap(~CpG,scales="free",ncol=2) + cowplot::theme_cowplot() +
  theme(legend.position = "none")

ggsave("figures/nonLinear_dDMPs_betavsstage.pdf", g, units="in", width=6, height=6, bg = "white")



```

```{r}

# Get sig results
## Non-treat filtered
probes.eby.bmicont.df <- GetLimmaResults(eby.bmicont.fit, bmicont.contMatrix, annotdata)

 
## Treat-filtered (abs log2FC > n)
probes.trt.bmicont.df <- GetLimmaResults(trt.bmicont.fit, bmicont.contMatrix, annotdata, trt_toggle=TRUE)


#add bonferroni - adjust per contrast
probes.eby.bmicont.df <- probes.eby.bmicont.df %>%
  group_by(CONTRAST) %>%
  mutate(bonferroni = p.adjust(P.Value,method="bonferroni")) %>% 
  as.data.frame()

probes.trt.bmicont.df <- probes.trt.bmicont.df %>%
  group_by(CONTRAST) %>%
  mutate(bonferroni = p.adjust(P.Value,method="bonferroni")) %>% 
  as.data.frame()
  
  




```

### Check n sig probes of interest

```{r}

nrow(probes.eby.bmicont.df[probes.eby.bmicont.df$bonferroni < 0.05 & !is.na(probes.eby.bmicont.df$adj.P.Val) & probes.eby.bmicont.df$CONTRAST == "Z_STAGE_WEEKS",])

nrow(probes.trt.bmicont.df[probes.trt.bmicont.df$bonferroni < 0.05 & !is.na(probes.trt.bmicont.df$adj.P.Val) & probes.trt.bmicont.df$CONTRAST == "Z_STAGE_WEEKS",])

nrow(probes.eby.bmicont.df[probes.eby.bmicont.df$bonferroni < 0.05 & !is.na(probes.eby.bmicont.df$adj.P.Val) & probes.eby.bmicont.df$CONTRAST == "sexmale - sexfemale",])

```
### Compare fun to qnt normalisation
```{r}
#get qnt m mat
qnt.m_mat <- minfi::getM(qnt.mSetSqFlt)
qnt.m_mat <- as.matrix(na.omit(qnt.m_mat))

bmicont.qnt.fit <- lmFit(qnt.m_mat[,rownames(bmicont.design)],bmicont.design)

# fit the contrasts
bmicont.qnt.fit <- contrasts.fit(bmicont.qnt.fit, bmicont.contMatrix)

# Treat variance (ie lfc > 0)
trt.bmicont.qnt.fit <- treat(bmicont.qnt.fit, lfc=trt_log2fc_thr)

#get limma res for qnt
probes.trt.bmicont.qnt.df <- GetLimmaResults(trt.bmicont.qnt.fit, bmicont.contMatrix, annotdata, trt_toggle=TRUE)

rm(qnt.m_mat)


probes.trt.bmicont.qnt.df <- probes.trt.bmicont.qnt.df %>%
  group_by(CONTRAST) %>%
  mutate(bonferroni = p.adjust(P.Value,method="bonferroni")) %>% 
  as.data.frame()

sigStageQnt <- probes.trt.bmicont.qnt.df[ probes.trt.bmicont.qnt.df$CONTRAST=="Z_STAGE_WEEKS" & probes.trt.bmicont.qnt.df$bonferroni <= 0.05,]


sigStageFun <- probes.trt.bmicont.df[ probes.trt.bmicont.df$CONTRAST=="Z_STAGE_WEEKS" & probes.trt.bmicont.df$bonferroni <= 0.05,]


g <- ggvenn::ggvenn(list("Functional"=sigStageFun$Name,"Quantile"=sigStageQnt$Name),show_percentage = FALSE)

ggsave(paste0(de_dir,"dDMPs_normalisation_comparison.png"), g, units="in", width=7, height=6, bg = "white")

ggsave("figures/dDMPs_normalisation_comparison.pdf", g, units="in", width=7, height=6, bg = "white")


```





### Annotate to genes

```{r}

# Get sig results
## Non-treat filtered

probes.eby.bmicont.df   <- cbind(probes.eby.bmicont.df, 
                                 annotateGenes(data.frame(chr=probes.eby.bmicont.df$chr, start=probes.eby.bmicont.df$pos, end=probes.eby.bmicont.df$pos))[,c("SYMBOL","DISTANCE")])


## Treat-filtered (abs log2FC > n)

probes.trt.bmicont.df   <- cbind(probes.trt.bmicont.df, 
                                 annotateGenes(data.frame(chr=probes.trt.bmicont.df$chr, start=probes.trt.bmicont.df$pos, end=probes.trt.bmicont.df$pos))[,c("SYMBOL","DISTANCE")])


#annotate the spline df
splineDF   <- cbind(splineDF, 
                                 annotateGenes(data.frame(chr=splineDF$chr, start=splineDF$pos, end=splineDF$pos))[,c("SYMBOL","DISTANCE")])


```

### Get mean sex CpG

```{r}

# Get Beta values
fun.beta_mat <- data.frame(minfi::getBeta(fun.mSetSqFlt))

# Get male/female samples
maleSamples <- as.character(metadata[metadata$sex == "male",]$Sample_Name)
femaleSamples <- as.character(metadata[metadata$sex == "female",]$Sample_Name)

# Make tall
tall_fun.beta_mat <- melt(cbind(CPG=rownames(fun.beta_mat), fun.beta_mat), id.vars="CPG")
tall_fun.beta_mat$variable <- gsub("X","",tall_fun.beta_mat$variable)

# aggregate mean CpG
meanMaleBeta   <- aggregate(. ~ CPG, 
                            data=tall_fun.beta_mat[which(tall_fun.beta_mat$variable %in% maleSamples),-2],
                            FUN=mean)
meanFemaleBeta <- aggregate(. ~ CPG, 
                            data=tall_fun.beta_mat[which(tall_fun.beta_mat$variable %in% femaleSamples),-2],
                            FUN=mean)
meanBeta <- cbind(meanMaleBeta, meanFemaleBeta[,-1,drop=FALSE])
colnames(meanBeta) <- c("CpG","MeanMale","MeanFemale")

# Combine
probes.eby.bmicont.df <- merge(probes.eby.bmicont.df, meanBeta, by.x="Name", by.y="CpG")
probes.trt.bmicont.df <- merge(probes.trt.bmicont.df, meanBeta, by.x="Name", by.y="CpG")

```

### Save

```{r}

# Save
write.table(probes.eby.bmicont.df, paste0(de_dir, "differential_methylation.limma_results.continuous_covariates.ebayes_adjusted.tsv"), sep="\t", quote=FALSE, row.names=FALSE)

write.table(probes.trt.bmicont.df, paste0(de_dir, "differential_methylation.limma_results.continuous_covariates.treat_log2FC",trt_log2fc_thr,"_thr.tsv"), sep="\t", quote=FALSE, row.names=FALSE)


# Save

write.table(probes.eby.bmicont.df[probes.eby.bmicont.df$bonferroni < 0.05 & !is.na(probes.eby.bmicont.df$adj.P.Val),], 
            paste0(de_dir, "differential_methylation.limma_results.continuous_covariates.ebayes_adjusted.padj_thr",0.05,".tsv"), sep="\t", quote=FALSE, row.names=FALSE)

write.table(probes.eby.bmicont.df[probes.eby.bmicont.df$bonferroni < 0.05 & !is.na(probes.eby.bmicont.df$adj.P.Val),], 
            paste0(de_dir, "differential_methylation.limma_results.continuous_covariates.ebayes_adjusted.padj_thr",0.05,".tsv"), sep="\t", quote=FALSE, row.names=FALSE)

write.table(probes.trt.bmicont.df[probes.trt.bmicont.df$bonferroni < 0.05 & !is.na(probes.trt.bmicont.df$adj.P.Val),], 
            paste0(de_dir, "differential_methylation.limma_results.continuous_covariates.treat_log2FC",trt_log2fc_thr,"_thr.padj_thr",0.05,".tsv"), sep="\t", quote=FALSE, row.names=FALSE)


write.table(splineDF_sig, paste0(de_dir, "differential_methylation.limma_results.spline.stage.tsv"), sep="\t", quote=FALSE, row.names=FALSE)

```

### Save desired

```{r}

# Save
## Tables - NOTE that these are for overlapping with DMRs
write.table(probes.eby.bmicont.df[probes.eby.bmicont.df$CONTRAST == "sexmale - sexfemale",], 
            paste0(de_dir, "differential_methylation.limma_results.sex.tsv"), sep="\t", quote=FALSE, row.names=FALSE)
write.table(probes.trt.bmicont.df[probes.trt.bmicont.df$CONTRAST == "Z_STAGE_WEEKS",], 
            paste0(de_dir, "differential_methylation.limma_results.stage_as_continuous.treat_log2FC",trt_log2fc_thr,"_thr.tsv"), sep="\t", quote=FALSE, row.names=FALSE)
write.table(probes.eby.bmicont.df[probes.eby.bmicont.df$CONTRAST == "sexmale - sexfemale" & probes.eby.bmicont.df$bonferroni < 0.05,], 
            paste0(de_dir, "differential_methylation.limma_results.sex.bonferonni",0.05,".tsv"), sep="\t", quote=FALSE, row.names=FALSE)
write.table(probes.trt.bmicont.df[probes.trt.bmicont.df$CONTRAST == "Z_STAGE_WEEKS" & probes.trt.bmicont.df$bonferroni < 0.05,],
            paste0(de_dir, "differential_methylation.limma_results.stage_as_continuous.treat_log2FC",trt_log2fc_thr,"_thr.bonferroni",0.05,".tsv"), sep="\t", quote=FALSE, row.names=FALSE)


## Bed
write.table(probes.eby.bmicont.df[probes.eby.bmicont.df$CONTRAST == "sexmale - sexfemale",c("chr","pos","pos","Name")], 
            paste0(de_dir, "differential_methylation.limma_results.sex.bed"), sep="\t", quote=FALSE, row.names=FALSE, col.names=FALSE)
write.table(probes.trt.bmicont.df[probes.trt.bmicont.df$CONTRAST == "Z_STAGE_WEEKS",c("chr","pos","pos","Name")], 
            paste0(de_dir, "differential_methylation.limma_results.stage_as_continuous.treat_log2FC",trt_log2fc_thr,"_thr.bed"), sep="\t", quote=FALSE, row.names=FALSE, col.names=FALSE)
write.table(probes.eby.bmicont.df[probes.eby.bmicont.df$CONTRAST == "sexmale - sexfemale"& probes.eby.bmicont.df$bonferroni < 0.05,c("chr","pos","pos","Name")], 
            paste0(de_dir, "differential_methylation.limma_results.sex.bonferroni",0.05,".bed"), sep="\t", quote=FALSE, row.names=FALSE, col.names=FALSE)
write.table(probes.trt.bmicont.df[probes.trt.bmicont.df$CONTRAST == "Z_STAGE_WEEKS"& probes.trt.bmicont.df$bonferroni < 0.05,c("chr","pos","pos","Name")], 
            paste0(de_dir, "differential_methylation.limma_results.stage_as_continuous.treat_log2FC",trt_log2fc_thr,"_thr.bonferroni",0.05,".bed"), sep="\t", quote=FALSE, row.names=FALSE, col.names=FALSE)

```

### Sex volcano plot for supplementry
```{r}

SexVolcanoPlot <- function(res, comparison, padj_thr=0.05, topn_genes=10, label_sig_thr=0.05, n.max_ovlp=25) {
  

    # Subset by contrast
    res_df <- res[res$CONTRAST == comparison,]
    
    # Remove NA
    res_df <- res_df[!is.na(res_df$bonferroni),]
    
    # Get sig
    sig.res_df <- res_df[res_df$bonferroni < padj_thr & !is.na(res_df$bonferroni),]
    
    # Add sig key
    res_df["Sig"] <- "Non-Sig"
    res_df[which(res_df$Name %in% sig.res_df$Name),]["Sig"] <- paste0("bonferroni<",padj_thr)
    
    # Add lfc key
    res_df["Direction"] <- "Up"
    res_df[res_df$logFC < 0,]["Direction"] <- "Down"
    
    # Add colour key
    res_df["ColorKey"] <- paste(res_df$Direction, res_df$Sig, sep=" ")
    res_df[grepl("Non-Sig",res_df$ColorKey),]["ColorKey"] <- "Non-Sig"
    
    # Order plotting
    res_df[ res_df$ColorKey==paste0("Down bonferroni<",padj_thr),"ColorKey"] <- "Bonferroni<0.05\n(higher DNAm in F)"
    res_df[ res_df$ColorKey==paste0("Up bonferroni<",padj_thr),"ColorKey"] <- "Bonferroni<0.05\n(higher DNAm in M)"
    res_df$ColorKey <- factor(res_df$ColorKey , levels=c("Non-Sig",  "Bonferroni<0.05\n(higher DNAm in F)", "Bonferroni<0.05\n(higher DNAm in M)"))
    
    # Order by LFCs (for labelling)
    res_df <- res_df[order(res_df$logFC, decreasing=TRUE),]
    
    # Define genes to label
    sig.res_df <- res_df[res_df$bonferroni < padj_thr & !is.na(res_df$bonferroni),]
    #label_df <- rbind(head(sig.res_df, topn_genes), tail(sig.res_df, topn_genes), sig.res_df[sig.res_df$padj < label_sig_thr,])
    sig.res_df$pi <- sig.res_df$logFC * -log10(sig.res_df$bonferroni)
    sig.res_df <- sig.res_df[order(sig.res_df$pi, decreasing=TRUE),]
    label_df <- rbind(head(sig.res_df, topn_genes), tail(sig.res_df, topn_genes),sig.res_df[sig.res_df$padj < label_sig_thr,])
    label_df <- label_df[!duplicated(label_df),]
    
    
    
    
    # Add size key
    res_df["SHAPE"] <- "Circle"
    res_df[which(res_df$SYMBOL %in% label_df$SYMBOL),]["SHAPE"] <- "Triangle"
    
    # Plot
    plt <- ggplot(res_df, aes(x=logFC, y=-log10(bonferroni), fill=ColorKey, color=ColorKey)) + 
      geom_vline(xintercept=0, linetype="dashed", color="darkgrey") +
      geom_point(alpha=0.75, shape=21, size=3) + 
      geom_text_repel(data=label_df, mapping=aes(x=logFC, y=-log10(bonferroni), label=SYMBOL), size=3.5,fontface ="italic",
                      show.legend = FALSE, min.segment.length = 0, max.overlaps=Inf,seed = 12345) +
      scale_fill_manual(values=c("Non-Sig"="lightgrey", "Bonferroni<0.05\n(higher DNAm in F)"="#BB5566", "Bonferroni<0.05\n(higher DNAm in M)"="#aed2e6")) +
      scale_color_manual(values=c("Non-Sig"="lightgrey","Bonferroni<0.05\n(higher DNAm in F)"="black", "Bonferroni<0.05\n(higher DNAm in M)"="black")) +
      geom_hline(yintercept=-log10(padj_thr), linetype="dashed", color="darkgrey") +
      ylab("-log10 Bonferroni") + xlab("Log2 FC")  + cowplot::theme_cowplot(font_size = 20) +
       theme(legend.title=element_blank(),legend.spacing.y = unit(3, 'cm')) +
       theme(legend.text=element_text(size=rel(0.5)))

    ggsave(paste0(de_dir,"differentialsexDMP_volcanoPlot.png"), plt, units="in", width=7, height=5, bg = "white")
    
        ggsave("figures/SuppFig_differentialsexDMP_volcanoPlot.pdf", plt, units="in", width=7, height=5, bg = "white")
  
}

SexVolcanoPlot(probes.eby.bmicont.df,comparison = "sexmale - sexfemale")



```

### Compare sex DMPs to previous study

```{r}
# blood <- readxl::read_xlsx("https://ndownloader.figstatic.com/files/42109380",sheet=2, skip=1)
blood <- readxl::read_xlsx("blood_sex_dmps.xlsx",sheet=2, skip=1)
sigSex <- probes.eby.bmicont.df[probes.eby.bmicont.df$CONTRAST == "sexmale - sexfemale" & probes.eby.bmicont.df$bonferroni < 0.05,]

#percentage overlap
table(sigSex$Name %in% blood$ID)/sum(table(sigSex$Name %in% blood$ID))

combined <-  merge(blood,sigSex,by.x="ID",by.y="Name")


label_df <- combined %>% dplyr::filter(dplyr::dense_rank(logFC.x) <= 5 | dplyr::dense_rank(-logFC.x) <= 5)

g <- ggplot(combined, aes(x = logFC.x,y = logFC.y)) +
geom_point()  +
geom_smooth(method = "lm", se=FALSE, color="red", formula = y ~ x) +
cowplot::theme_cowplot(font_size = 20) +
xlab("Blood Sex DMP effect size (log2 FC)") +
ylab("Cartilage Sex DMP effect size (log2 FC)") +
stat_cor(size=5) +
  geom_text_repel(data=label_df, mapping=aes(x=logFC.x, y=logFC.y, label=SYMBOL), size=4,fontface ="italic",
                      show.legend = FALSE, min.segment.length = 0, max.overlaps=Inf,seed = 12345) 

 ggsave(paste0(de_dir,"differentialsexDMP_blood_cartilage_effectSize.png"), g, units="in", width=7, height=6, bg = "white")
 
  ggsave("figures/SuppFig_differentialsexDMP_blood_cartilage_effectSize.pdf", g, units="in", width=7, height=6, bg = "white")



```

### Violin plots of top sDMPs

```{r}
topSigSexDMPs <- sigSex %>%  dplyr::slice_min(bonferroni,n=20) %>% dplyr::pull(Name)



sexBeta <- fun.beta_mat[topSigSexDMPs, ]
sexBeta <- tibble::rownames_to_column(sexBeta,"CpG")
sexBeta <-reshape2::melt(sexBeta)
sexBeta$sex <- ifelse(sexBeta$variable %in% paste0("X",maleSamples),"Male","Female")
sexBeta$CpG <- factor(sexBeta$CpG,levels=topSigSexDMPs)

g <- ggplot(sexBeta, aes(x = CpG, y = value, fill = sex)) +
    geom_violin(draw_quantiles = 0.5, width=2, position=position_dodge(width=0.75)) +
    labs(y = "Methylation (Beta)",
         x = "") +
    scale_fill_manual(values = c("Male" = "#BB5566", "Female" = "#aed2e6")) +
    cowplot::theme_cowplot(font_size = 16) +
  theme(axis.text.x = element_text(angle = 45, hjust = 1)) +
    geom_vline(xintercept = seq(0.5, length(unique(sexBeta$CpG)) - 0.5), linetype = "dotted", color = "gray")


 ggsave(paste0(de_dir,"differentialsexDMP_violinPlot.png"), g, units="in", width=10, height=5, bg = "white")
 
 
  ggsave("figures/SuppFig_differentialsexDMP_violinPlot.pdf", g, units="in", width=10, height=5, bg = "white")

```




### Volcano plots

```{r}

# Plot nice volcano plots
NiceVolcanoPlots(probes.eby.bmicont.df, comparison="CONTRAST", dir=de_dir, pval_colname="bonferroni", logfc_colname="logFC", feature_colname="Name", 
                 padj_thr=0.05, label_sig_thr=1e-100, topn_genes=0, logfc_thr=0.1, fills=c(NA,NA,NA), colours=c("lightgrey", "#BB5566", "#004488"), extra_filename="nonlogFC_thresholded")

# Plot nice volcano plots
NiceVolcanoPlots(probes.trt.bmicont.df, comparison="CONTRAST", dir=de_dir, pval_colname="bonferroni", logfc_colname="logFC", feature_colname="Name", 
                 padj_thr=0.05, label_sig_thr=1e-100, topn_genes=0, logfc_thr=0.1,
                 fills=c(NA,NA,NA), colours=c("lightgrey", "#BB5566", "#004488"), extra_filename=paste0("limmatreat_logFC_thr",trt_log2fc_thr))

```

### Relationship between top CpGs and stage

```{r}

TopFeaturesLinePlot <- function(sig_df, meta, beta_vals, n_cpgs=10, fileSuffix="", order_by="logFC", extra=xlab("Value"), out_dir="./") {
  
  # Dependencies
  require(ggplot2)
  
  # Order by order column
  sig_df <- sig_df[order(sig_df[,order_by], decreasing=TRUE),]
  
  # Get top/bottom n cpgs by logFC
  top.pos_cpgs <- sig_df[1:n_cpgs,]
  top.neg_cpgs <- sig_df[(nrow(sig_df)-(1:n_cpgs)),]
  
  # Split up/down-
  top.pos_cpgs <- top.pos_cpgs[top.pos_cpgs[,order_by] > 0,]$Name
  top.neg_cpgs <- top.neg_cpgs[top.neg_cpgs[,order_by] < 0,]$Name
  
  # Fix matrix
  colnames(beta_vals) <- gsub("^X","",colnames(beta_vals))
  
  # Get metadata and methylation_data
  meta$Sample_Name <- as.character(meta$Sample_Name)
  top.combined_data.df <- cbind(meta, t(beta_vals[top.pos_cpgs,meta$Sample_Name]))
  bot.combined_data.df <- cbind(meta, t(beta_vals[top.neg_cpgs,meta$Sample_Name]))
  
  # Add output
  out_dir <- paste0(out_dir, "/topCpGs_by",order_by,"/")
  dir.create(out_dir, recursive=TRUE)
  
  # Loop over cpgs
  for (cpg in top.pos_cpgs) { 
    
    # Plot
    plt <- ggplot(top.combined_data.df, aes(x=stage_weeks, y=.data[[cpg]])) + 
      geom_point(shape=21, fill="darkred", alpha=0.65, size=5) + 
      ggtitle(cpg) +
      extra +
      theme_bw(base_size=16) + theme(axis.title=element_text(size=24),
                                     plot.title=element_text(size=30))
      
    # Save
    ggsave(paste0(out_dir, cpg, ".orderedBy_top",order_by,fileSuffix,".linePlot.png"), units="px", width=4000, height=3000)
    
  }
  
  # Loop over cpgs
  for (cpg in top.neg_cpgs) {
    
    # Plot
    plt <- ggplot(bot.combined_data.df, aes(x=stage_weeks, y=.data[[cpg]])) + 
      geom_point(shape=21, fill="darkblue", alpha=0.65, size=5) + 
      ggtitle(cpg) +
      extra + 
      theme_bw(base_size=16) + theme(axis.title=element_text(size=24),
                                     plot.title=element_text(size=30))
      
    # Save
    ggsave(paste0(out_dir, cpg, ".orderedBy_bottom",order_by,fileSuffix,".linePlot.png"), units="px", width=4000, height=3000)
    
  }

}  

# Plot
TopFeaturesLinePlot(sig_df=probes.trt.bmicont.df[probes.trt.bmicont.df$bonferroni < 0.05 & 
                                                 !is.na(probes.trt.bmicont.df$adj.P.Val) & 
                                                 probes.trt.bmicont.df$CONTRAST == "Z_STAGE_WEEKS",],
                    meta=for_de.metadata, beta_vals=fun.beta_mat, n_cpgs=25, fileSuffix=".methylation_beta_vs_stage",
                    extra=labs(x="Stage (Weeks)", y="Beta"), out_dir=de_dir)

```

## DMR work

### Dmrff

Based on \* http://htmlpreview.github.io/?https://github.com/perishky/dmrff/blob/master/docs/tutorial.html \* http://htmlpreview.github.io/?https://github.com/perishky/dmrff/blob/master/docs/correlated-cpg-sites.html

#### Correlations between CpGs

Regions that fail Dmrff tend to comprise CpGs with strong dependencies on each-other, ie any one CpG explains a considerable proportion of the variance, or the association between the region and variable would tend to disappear when one or a few CpGs are included in the model. Therefore the region doesn't explain much more variance than any individual CpG. Conversely CpGs that do pass tend to explain different proportions of variance to the variable of interest.


#### Run DMRFF
 
```{r}

# Define gap vals
gapvals <- c(1000)

# Loop over max gap values and run Dmrff on limma output
annotdata <- annotdata[ annotdata$Name %in% rownames(fun.m_mat),]

# Split into up- and down-reg


## Treat
# trt.sex.dmrff_df        <- DmrffTuneMaxGap(fit=trt.bmicont.fit, annotation=annotdata, m=fun.m_mat, all_contrasts=c("sexmale"),
#                                             extra=paste0(".groupedby_sex.treat_log2FC",trt_log2fc_thr), sdir=dmr_dir, max_gap_vals=gapvals)
#trt.bmigroup.dmrff_df   <- DmrffTuneMaxGap(fit=trt.bmigroup.fit, annotation=annotdata, m=fun.m_mat, all_contrasts=colnames(bmigroup.contMatrix), 
#                                            extra=paste0(".groupedby_bmi_by_group.treat_log2FC",trt_log2fc_thr), dir=dmr_dir, max_gap_vals=gapvals)
trt.stgweek.dmrff_df    <- DmrffTuneMaxGap(fit=trt.bmicont.fit, annotation=annotdata, m=fun.m_mat, all_contrasts=c("Z_STAGE_WEEKS"), reGenerateFile=TRUE,
                                            extra=paste0(".groupedby_stage_as_continuous.treat_log2FC",trt_log2fc_thr), dir=dmr_dir, max_gap_vals=gapvals)
#trt.sexstgweek.dmrff_df <- DmrffTuneMaxGap(fit=trt.bmicont.fit, annotation=annotdata, m=fun.m_mat, all_contrasts=c("sexmale_Z_STAGE_WEEKS"),
#                                            extra=paste0(".groupedby_sex_stage_as_continuous.treat_log2FC",trt_log2fc_thr), dir=dmr_dir, max_gap_vals=gapvals)
#trt.bmicont.dmrff_df    <- DmrffTuneMaxGap(fit=trt.bmicont.fit, annotation=annotdata, m=fun.m_mat, all_contrasts=c("Z_MAT_BMI"),
#                                            extra=paste0(".groupedby_bmi_as_continuous.treat_log2FC",trt_log2fc_thr), dir=dmr_dir, max_gap_vals=gapvals)
#trt.bmibinary.dmrff_df  <- DmrffTuneMaxGap(fit=trt.bmibinary.fit, annotation=annotdata, m=fun.m_mat, all_contrasts=colnames(bmibinary.contMatrix), 
#                                            extra=paste0(".groupedby_binarised_bmi.treat_log2FC",trt_log2fc_thr), dir=dmr_dir, max_gap_vals=gapvals)

## Ebayes
#eby.bmigroup.dmrff_df   <- DmrffTuneMaxGap(fit=eby.bmigroup.fit, annotation=annotdata, m=fun.m_mat, all_contrasts=colnames(bmigroup.contMatrix), 
#                                            extra=".groupedby_bmi_by_group", dir=dmr_dir, max_gap_vals=gapvals)
#eby.bmicont.dmrff_df    <- DmrffTuneMaxGap(fit=eby.bmicont.fit, annotation=annotdata, m=fun.m_mat, all_contrasts=c("Z_MAT_BMI"),
#                                            extra=".groupedby_bmi_as_continuous", dir=dmr_dir, max_gap_vals=gapvals)
# eby.stgweek.dmrff_df    <- DmrffTuneMaxGap(fit=eby.bmicont.fit, annotation=annotdata, m=fun.m_mat, all_contrasts=c("Z_STAGE_WEEKS"),
#                                             extra=".groupedby_stage_as_continuous", dir=dmr_dir, max_gap_vals=gapvals)
eby.sex.dmrff_df        <- DmrffTuneMaxGap(fit=eby.bmicont.fit, annotation=annotdata, m=fun.m_mat, all_contrasts=c("sexmale - sexfemale"), reGenerateFile=TRUE,
                                            extra=".groupedby_sex", dir=dmr_dir, max_gap_vals=gapvals)
#eby.sexstgweek.dmrff_df <- DmrffTuneMaxGap(fit=eby.bmicont.fit, annotation=annotdata, m=fun.m_mat, all_contrasts=c("sexmale_Z_STAGE_WEEKS"),
#                                            extra=".groupedby_sex_stage_as_continuous", dir=dmr_dir, max_gap_vals=gapvals)
#eby.bmibinary.dmrff_df  <- DmrffTuneMaxGap(fit=eby.bmibinary.fit, annotation=annotdata, m=fun.m_mat, all_contrasts=colnames(bmibinary.contMatrix), 
#                                            extra=".groupedby_binarised_bmi", dir=dmr_dir, max_gap_vals=gapvals)


```

#### Fix zero p-value

One DMR has a p-value lower than what my machine can handle so it defaults to 0. So we deal with it manually.

```{r}
pvalue.extreme <- function(z) {
  
   # Function adapted from:
   # https://stackoverflow.com/questions/46416027/how-to-compute-p-values-from-z-scores-in-r-when-the-z-score-is-large-pvalue-muc/46416222#46416222
  
   log.pvalue <- log(2) + pnorm(abs(z), lower.tail = FALSE, log.p = TRUE)
   log10.pvalue <- log.pvalue/log(10) ## from natural log to log10
   
   # Return
   return(log10.pvalue)
   
}


# Fix zero p-values (Just one for a redundant DMR)
log10_pval <- pvalue.extreme(trt.stgweek.dmrff_df[trt.stgweek.dmrff_df$p.adjust == 0,]$z)
trt.stgweek.dmrff_df[trt.stgweek.dmrff_df$p.value == 0,]["p.value"] <- 10^log10_pval
trt.stgweek.dmrff_df[trt.stgweek.dmrff_df$p.adjust == 0,]["p.adjust"] <- 10^log10_pval

```

#### Save

```{r}

# Save
write.table(trt.stgweek.dmrff_df, paste0(dmr_dir,"dmrff_results.stage_as_continuous.treat_log2FC",trt_log2fc_thr,"unfiltered.tsv"), sep="\t", quote=FALSE, row.names=FALSE)
write.table(eby.sex.dmrff_df, paste0(dmr_dir,"dmrff_results.sex.unfiltered.tsv"), sep="\t", quote=FALSE, row.names=FALSE)

```

#### Remove single CpGs

```{r}

# Drop single CpGs
## eBayes
eby.sex.dmrff_df <- eby.sex.dmrff_df[eby.sex.dmrff_df$n > 1,]

## Treat
trt.stgweek.dmrff_df <- trt.stgweek.dmrff_df[trt.stgweek.dmrff_df$n > 1,]


```

#### Visualise distributions of gaps

# Loop over gaps

for (gap in unique(trt.stgweek.dmrff_df$MAXGAP)) { ggsave(paste0(dmr_dir, "stage.number_of_cpgs.per_dmr.maxgap",gap,".png"), units="px", width=2500, height=2000, ggplot(trt.stgweek.dmrff_df[trt.stgweek.dmrff_df$MAXGAP == gap & trt.stgweek.dmrff_df$p.adjust < 0.001,], aes(x=n)) + geom_histogram(fill="darkorchid4", color="black", alpha=0.25) + ylim(0,12500) + xlim(0, 100) + geom_vline(xintercept=0, linetype="dashed",color="darkgrey") + geom_hline(yintercept=0, linetype="dashed",color="darkgrey") + labs(title=paste0("Max Gap = ",gap), x="Number of CpGs", y="Frequency") + theme_bw(base_size=16) + theme(axis.title=element_text(size=24))) } for (gap in unique(eby.sex.dmrff_df$MAXGAP)) { ggsave(paste0(dmr_dir, "sex.number_of_cpgs.per_dmr.maxgap",gap,".png"), units="px", width=2500, height=2000, ggplot(eby.sex.dmrff_df\[eby.sex.dmrff_df$MAXGAP == gap & eby.sex.dmrff_df$p.adjust \< 0.05,\], aes(x=n)) + geom_histogram(fill="darkorchid4", color="black", alpha=0.25) + ylim(0,50) + xlim(0, 20) + geom_vline(xintercept=0, linetype="dashed",color="darkgrey") + geom_hline(yintercept=0, linetype="dashed",color="darkgrey") + labs(title=paste0("Max Gap =",gap), x="Number of CpGs", y="Frequency") + theme_bw(base_size=16) + theme(axis.title=element_text(size=24))) }

```{r}

# Filter only DMRs equal to chosen max gap

trt.stgweek.dmrff_df  <-   trt.stgweek.dmrff_df[trt.stgweek.dmrff_df$MAXGAP == maxgap_val,]
eby.sex.dmrff_df      <-           eby.sex.dmrff_df[eby.sex.dmrff_df$MAXGAP == maxgap_val,]

# Sig filter
# trt.stgweek.dmrff_df <- trt.stgweek.dmrff_df[trt.stgweek.dmrff_df$p.adjust < 0.001 & !is.na(trt.stgweek.dmrff_df$chr),]
# trt.sex.dmrff_df <- trt.sex.dmrff_df[trt.sex.dmrff_df$p.adjust < 0.05 & !is.na(trt.sex.dmrff_df$chr),]

```

#### Add DMR IDs and CpGs

```{r}
# Add DMR annotations
gc()
possible_cpgs <- unique(probes.trt.bmicont.df[probes.trt.bmicont.df$CONTRAST == "Z_STAGE_WEEKS",]$Name)
possible_cpgannots <- annotdata[which(annotdata$Name %in% possible_cpgs),]
trt.stgweek.dmrff_df <- AddDMRCpGDetails(trt.stgweek.dmrff_df, 
                                         probes.trt.bmicont.df[probes.trt.bmicont.df$bonferroni < 0.05 & 
                                                               probes.trt.bmicont.df$CONTRAST == "Z_STAGE_WEEKS",], 
                                         possible_cpgannots)
possible_cpgs <- unique(probes.eby.bmicont.df[probes.eby.bmicont.df$CONTRAST == "sexmale - sexfemale",]$Name)
possible_cpgannots <- annotdata[which(annotdata$Name %in% possible_cpgs),]
eby.sex.dmrff_df     <- AddDMRCpGDetails(eby.sex.dmrff_df, 
                                         probes.eby.bmicont.df[probes.eby.bmicont.df$bonferroni < 0.05 & 
                                                               probes.eby.bmicont.df$CONTRAST == "sexmale - sexfemale",], 
                                         possible_cpgannots)

# Filter DMRs with NO CpGs. These are DMRs with only non-significant DMRs present
# Also drop single CpG DMRs
trt.stgweek.dmrff_df <- trt.stgweek.dmrff_df[trt.stgweek.dmrff_df$CpG != "" & trt.stgweek.dmrff_df$n > 1,]
eby.sex.dmrff_df <- eby.sex.dmrff_df[eby.sex.dmrff_df$CpG != "" & eby.sex.dmrff_df$n > 1,]

```

### Make tall format

```{r}

# Update DMR ID
trt.stgweek.dmrff_df["DMR_ID"] <- paste0("dDMR_",1:nrow(trt.stgweek.dmrff_df))
eby.sex.dmrff_df["DMR_ID"] <- paste0("sDMR_",1:nrow(eby.sex.dmrff_df))
rownames(trt.stgweek.dmrff_df) <- trt.stgweek.dmrff_df$DMR_ID
rownames(eby.sex.dmrff_df) <- eby.sex.dmrff_df$DMR_ID

# Get tall stage and sex dataframes
stageTall_df <- GetTall_DMR_DF(trt.stgweek.dmrff_df)
sexTall_df   <- GetTall_DMR_DF(eby.sex.dmrff_df)


```

### Check direction of CpGs

Note we accept CpGs with opposite effect sizes in DMRs as these have nominal p \< 0.05.

```{r}

# Add CpG log2FC
colnames(probes.trt.bmicont.df)[colnames(probes.trt.bmicont.df)=="bonferroni"] <- "probe_bonferroni"
colnames(probes.eby.bmicont.df)[colnames(probes.eby.bmicont.df)=="bonferroni"] <- "probe_bonferroni"

stageTall_df <- merge(stageTall_df, probes.trt.bmicont.df[probes.trt.bmicont.df$CONTRAST == "Z_STAGE_WEEKS",c("Name","logFC","pos","P.Value","adj.P.Val","probe_bonferroni")], by.x="CpG", by.y="Name")
sexTall_df <- merge(sexTall_df, probes.eby.bmicont.df[probes.eby.bmicont.df$CONTRAST == "sexmale - sexfemale",c("Name","logFC","pos","P.Value","adj.P.Val","probe_bonferroni")], by.x="CpG", by.y="Name")
stageTall_df <- stageTall_df[order(paste0(as.numeric(gsub("chr","",stageTall_df$chr)),stageTall_df$pos), decreasing=FALSE),]
sexTall_df <- sexTall_df[order(paste0(as.numeric(gsub("chr","",sexTall_df$chr)),sexTall_df$pos), decreasing=FALSE),]


```

### Plot

```{r}

# Get sig
sig_trt.stgweek.dmrff_df <- trt.stgweek.dmrff_df[trt.stgweek.dmrff_df$p.adjust < 0.05,]
sig_eby.sex.dmrff_df <- eby.sex.dmrff_df[eby.sex.dmrff_df$p.adjust < 0.05,]

# Order by n
sig_trt.stgweek.dmrff_df <- sig_trt.stgweek.dmrff_df[order(sig_trt.stgweek.dmrff_df$n, decreasing=TRUE),]
sig_eby.sex.dmrff_df <- sig_eby.sex.dmrff_df[order(sig_eby.sex.dmrff_df$n, decreasing=TRUE),]
rownames(sig_trt.stgweek.dmrff_df) <- sig_trt.stgweek.dmrff_df$DMR_ID
rownames(sig_eby.sex.dmrff_df) <- sig_eby.sex.dmrff_df$DMR_ID

# Label genes in top 25 n
trt.stgweek.dmrff_df["LABEL"] <- "ns"; trt.stgweek.dmrff_df[sig_trt.stgweek.dmrff_df$DMR_ID[1:20],]["LABEL"] <- "Top 15\nby N CpG"
eby.sex.dmrff_df["LABEL"] <- "ns"; eby.sex.dmrff_df[sig_eby.sex.dmrff_df$DMR_ID[1:15],]["LABEL"] <- "Top 15\nby N CpG"

# Plot
stg_plt <- ggplot(trt.stgweek.dmrff_df, aes(x=n, y=estimate, label=SYMBOL, fill=LABEL, color=LABEL)) + 
  geom_hline(yintercept=0, linetype="dashed", color="darkgrey") +
  geom_vline(xintercept=0, linetype="dashed", color="darkgrey") +
  geom_point(data=trt.stgweek.dmrff_df[trt.stgweek.dmrff_df$LABEL == "ns",], 
             mapping=aes(size=-log10(p.adjust)), shape=21, alpha=0.65) +
  geom_point(data=trt.stgweek.dmrff_df[trt.stgweek.dmrff_df$LABEL != "ns",], 
             mapping=aes(size=-log10(p.adjust)), shape=21, alpha=0.65) +
  scale_fill_manual(values=c("darkgrey", "red")) + scale_color_manual(values=c("darkgrey", "darkred")) + 
  geom_text_repel(data=trt.stgweek.dmrff_df[trt.stgweek.dmrff_df$LABEL != "ns",], 
                  color="darkred",size=3.5, force_pull=0.5) +
  ylim(-1.15, 1.15) +
  xlab("Number of CpGs") + ylab("Estimate") + ggtitle("Gestation Stage DMRs") +
  theme_bw(base_size=12) + theme(axis.title=element_text(size=16), plot.title=element_text(size=20)) +
  guides(fill=guide_legend(title=""), color=guide_legend(title=""))
sex_plt <- ggplot(eby.sex.dmrff_df, aes(x=n, y=estimate, label=SYMBOL, fill=LABEL, color=LABEL)) + 
  geom_hline(yintercept=0, linetype="dashed", color="darkgrey") +
  geom_vline(xintercept=0, linetype="dashed", color="darkgrey") +
  geom_point(data=eby.sex.dmrff_df[eby.sex.dmrff_df$LABEL == "ns",], 
             mapping=aes(size=-log10(p.adjust)), shape=21, alpha=0.65) +
  geom_point(data=eby.sex.dmrff_df[eby.sex.dmrff_df$LABEL != "ns",], 
             mapping=aes(size=-log10(p.adjust)), shape=21, alpha=0.65) +
  scale_fill_manual(values=c("darkgrey", "red")) + scale_color_manual(values=c("darkgrey", "darkred")) + 
  geom_text_repel(data=eby.sex.dmrff_df[eby.sex.dmrff_df$LABEL != "ns",], 
                  color="darkred",size=3.5, force_pull=2) +
  ylim(-2, 2) +
  xlab("Number of CpGs") + ylab("Estimate") + ggtitle("Sex DMRs") +
  theme_bw(base_size=12) + theme(axis.title=element_text(size=16), plot.title=element_text(size=20)) +
  guides(fill=guide_legend(title=""), color=guide_legend(title=""))

# Save
ggsave(paste0(dmr_dir, "dmrStage_nCpGVsEstimate.topLabelled.top15nCpG.pdf"), stg_plt, units="in", width=8, height=5, device="pdf", dpi=300)
ggsave(paste0(dmr_dir, "dmrSex_nCpGVsEstimate.topLabelled.top15nCpG.pdf"), sex_plt, units="px", width=3000, height=2250, device="pdf", dpi=300)

# Plot
trt.stgweek.dmrff_df$LABEL <- factor(trt.stgweek.dmrff_df$LABEL, levels=c("Top 15\nby N CpG","ns"))
eby.sex.dmrff_df$LABEL <- factor(eby.sex.dmrff_df$LABEL, levels=c("Top 15\nby N CpG","ns"))
stg_plt <- ggplot(trt.stgweek.dmrff_df, aes(label=SYMBOL, 
                                            fill=LABEL, 
                                            color=LABEL)) + 
  geom_hline(yintercept=0, linetype="dashed", color="darkgrey") +
  geom_vline(xintercept=0, linetype="dashed", color="darkgrey") +
  geom_point(data=trt.stgweek.dmrff_df[trt.stgweek.dmrff_df$LABEL == "ns",], 
             mapping=aes(y=-log10(p.adjust), x=estimate, size=n), shape=21, alpha=0.65) +
  geom_point(data=trt.stgweek.dmrff_df[trt.stgweek.dmrff_df$LABEL != "ns",], 
             mapping=aes(y=-log10(p.adjust), x=estimate, size=n), shape=21, alpha=0.65) +
  scale_fill_manual(values=c("Top 15\nby N CpG"="red", "ns"="darkgrey")) + 
  scale_color_manual(values=c("Top 15\nby N CpG"="darkred", "ns"="darkgrey")) +
  geom_text_repel(data=trt.stgweek.dmrff_df[trt.stgweek.dmrff_df$LABEL != "ns",], 
                  mapping=aes(y=-log10(p.adjust), x=estimate),
                  color="darkred",size=3.5, force_pull=0.5) +
  ylim(0, 325) +
  xlab("Estimate") + ylab("-Log10 Adjusted P-Value") + ggtitle("Gestation Stage DMRs") +
  theme_bw(base_size=12) + theme(axis.title=element_text(size=16), plot.title=element_text(size=20)) +
  guides(fill=guide_legend(title=""), color=guide_legend(title=""), size=guide_legend(title="N CpGs"))
sex_plt <- ggplot(eby.sex.dmrff_df, aes(label=SYMBOL, 
                                            fill=factor(LABEL, levels=c("Top 15\nby N CpG","ns")), 
                                            color=factor(LABEL, levels=c("Top 15\nby N CpG","ns")))) + 
  geom_hline(yintercept=0, linetype="dashed", color="darkgrey") +
  geom_vline(xintercept=0, linetype="dashed", color="darkgrey") +
  geom_point(data=eby.sex.dmrff_df[eby.sex.dmrff_df$LABEL == "ns",], 
             mapping=aes(y=-log10(p.adjust), x=estimate, size=n), shape=21, alpha=0.65) +
  geom_point(data=eby.sex.dmrff_df[eby.sex.dmrff_df$LABEL != "ns",], 
             mapping=aes(y=-log10(p.adjust), x=estimate, size=n), shape=21, alpha=0.65) +
  scale_fill_manual(values=c("Top 15\nby N CpG"="red", "ns"="darkgrey")) + 
  scale_color_manual(values=c("Top 15\nby N CpG"="darkred", "ns"="darkgrey")) +
  geom_text_repel(data=eby.sex.dmrff_df[eby.sex.dmrff_df$LABEL != "ns",], 
                  mapping=aes(y=-log10(p.adjust), x=estimate),
                  color="darkred",size=1.5, force=25) +
  xlab("Estimate") + ylab("-Log10 Adjusted P-Value") + ggtitle("Sex DMRs") +
  theme_bw(base_size=12) + theme(axis.title=element_text(size=16), plot.title=element_text(size=20)) +
  guides(fill=guide_legend(title=""), color=guide_legend(title=""), size=guide_legend(title="N CpGs"))

# Save
ggsave(paste0(dmr_dir, "dmrStage_EstimateVspadj.topLabelled.top15nCpG.pdf"), stg_plt, units="px", width=2250, height=2250, device="pdf", dpi=300)
ggsave(paste0(dmr_dir, "dmrSex_EstimateVspadj.topLabelled.top15nCpG.pdf"), sex_plt, units="px", width=2250, height=2250, device="pdf", dpi=300)

```

### Save DMRs

```{r}

# Save
write.table(trt.stgweek.dmrff_df[,
            c("DMR_ID","SITE","chr","start","end","SYMBOL","DISTANCE","B","S","estimate","se","z","p.value","p.adjust","CONTRAST","MAXGAP","CpG","n","nSigCpg","nNonSigCpg","pSigCpg")], 
            file=paste0(dmr_dir, "dmrff_results.stage_as_continuous.treat_log2FC",trt_log2fc_thr,".tsv"),
            col.names = TRUE,row.names = FALSE, sep = "\t", quote = FALSE)
# Save
write.table(eby.sex.dmrff_df[,
            c("DMR_ID","SITE","chr","start","end","SYMBOL","DISTANCE","B","S","estimate","se","z","p.value","p.adjust","CONTRAST","MAXGAP","CpG","n","nSigCpg","nNonSigCpg","pSigCpg")], 
            file=paste0(dmr_dir, "dmrff_results.sex.tsv"),
            col.names = TRUE,row.names = FALSE, sep = "\t", quote = FALSE)

# Save
write.table(trt.stgweek.dmrff_df[trt.stgweek.dmrff_df$estimate > 0,
            c("DMR_ID","SITE","chr","start","end","SYMBOL","DISTANCE","B","S","estimate","se","z","p.value","p.adjust","CONTRAST","MAXGAP","CpG","n","nSigCpg","nNonSigCpg","pSigCpg")], 
            file=paste0(dmr_dir, "dmrff_results.stage_as_continuous.treat_log2FC",trt_log2fc_thr,".up.tsv"),
            col.names = TRUE,row.names = FALSE, sep = "\t", quote = FALSE)
write.table(trt.stgweek.dmrff_df[trt.stgweek.dmrff_df$estimate < 0, 
            c("DMR_ID","SITE","chr","start","end","SYMBOL","DISTANCE","B","S","estimate","se","z","p.value","p.adjust","CONTRAST","MAXGAP","CpG","n","nSigCpg","nNonSigCpg","pSigCpg")], 
            file=paste0(dmr_dir, "dmrff_results.stage_as_continuous.treat_log2FC",trt_log2fc_thr,".down.tsv"),
            col.names = TRUE,row.names = FALSE, sep = "\t", quote = FALSE)
write.table(eby.sex.dmrff_df[eby.sex.dmrff_df$estimate > 0,
            c("DMR_ID","SITE","chr","start","end","SYMBOL","DISTANCE","B","S","estimate","se","z","p.value","p.adjust","CONTRAST","MAXGAP","CpG","n","nSigCpg","nNonSigCpg","pSigCpg")], 
            file=paste0(dmr_dir, "dmrff_results.sex.up.tsv"),
            col.names = TRUE,row.names = FALSE, sep = "\t", quote = FALSE)
write.table(eby.sex.dmrff_df[eby.sex.dmrff_df$estimate < 0,
            c("DMR_ID","SITE","chr","start","end","SYMBOL","DISTANCE","B","S","estimate","se","z","p.value","p.adjust","CONTRAST","MAXGAP","CpG","n","nSigCpg","nNonSigCpg","pSigCpg")], 
            file=paste0(dmr_dir, "dmrff_results.sex.down.tsv"),
            col.names = TRUE,row.names = FALSE, sep = "\t", quote = FALSE)

# Save
write.table(stageTall_df[,
            c("DMR_ID","SITE","chr","start","end","SYMBOL","DISTANCE","B","S","estimate","se","z","p.value","p.adjust","CONTRAST","MAXGAP","CpG","pos","logFC","P.Value","adj.P.Val","probe_bonferroni","n","nSigCpg","nNonSigCpg","pSigCpg")], 
            file=paste0(dmr_dir, "dmrff_results.stage_as_continuous.treat_log2FC",trt_log2fc_thr,".preClust_TallFormat.tsv"),
            col.names = TRUE,row.names = FALSE, sep = "\t", quote = FALSE)
# Save
write.table(sexTall_df[,
            c("DMR_ID","SITE","chr","start","end","SYMBOL","DISTANCE","B","S","estimate","se","z","p.value","p.adjust","CONTRAST","MAXGAP","CpG","pos","logFC","P.Value","adj.P.Val","probe_bonferroni","n","nSigCpg","nNonSigCpg","pSigCpg")], 
            file=paste0(dmr_dir, "dmrff_results.sex.preClust_TallFormat.tsv"),
            col.names = TRUE,row.names = FALSE, sep = "\t", quote = FALSE)

```

### Add mean DMR M-values

```{r}

# Add DMR IDs to M value matrix
mMat_byStageDMR <- merge(stageTall_df[,c("CpG","DMR_ID")],fun.m_mat, by.x="CpG", by.y="row.names")
mMat_bySexDMR <- merge(sexTall_df[,c("CpG","DMR_ID")], fun.m_mat, by.x="CpG", by.y="row.names")

# Aggregate to mean
mMatMean_byStageDMR <- aggregate(. ~ DMR_ID, data=mMat_byStageDMR[,-1], FUN=mean)
mMatMean_bySexDMR <- aggregate(. ~ DMR_ID, data=mMat_bySexDMR[,-1], FUN=mean)

# Save
write.table(mMat_byStageDMR, paste0(dmr_dir,"dmrff_dmrs.stage_as_continuous.mean_cpg_mval.maxGap",maxgap_val,"_withIDs.tsv"), quote=FALSE, sep="\t")
write.table(mMatMean_bySexDMR, paste0(dmr_dir,"dmrff_dmrs.sex.mean_cpg_mval.maxGap",maxgap_val,"_withIDs.tsv"), quote=FALSE, sep="\t")

```

### Get DMR CpG methylation

```{r}

# Save
write.table(cbind(CpG=stageTall_df$CpG, 
                  SigCpG=ifelse(stageTall_df$P.Value < 0.05, "(nominal) p<0.05", "ns"), 
                  SigDMR=ifelse(stageTall_df$p.adjust < 0.05, "(DMR) bonferonni<0.05", "ns"), 
                  data.frame(fun.m_mat[stageTall_df$CpG,])), 
            paste0(dmr_dir, "dmrff_results.stage_as_continuous.treat_log2FC",trt_log2fc_thr,".cpgMvals.tsv"),
            sep="\t", quote=FALSE, col.names=TRUE, row.names=FALSE)
write.table(cbind(CpG=sexTall_df$CpG, 
                  SigCpG=ifelse(sexTall_df$P.Value < 0.05, "(nominal) p<0.05", "ns"), 
                  SigDMR=ifelse(sexTall_df$p.adjust < 0.05, "(DMR) bonferonni<0.05", "ns"), 
                  data.frame(fun.m_mat[sexTall_df$CpG,])), 
            paste0(dmr_dir, "dmrff_results.sex.cpgMvals.tsv"),
            sep="\t", quote=FALSE, col.names=TRUE, row.names=FALSE)

```

### Gene ontology

For region, background is all CpGs

#### CpG

```{r}

# sig_cpgs <- c("Z_STAGE_WEEKS"=0.001, "sexmale - sexfemale"=0.05)

sig_cpgs <- c("Stage_spline"=0.05)

# Split by up- and down-regulation
for (direction in c("up","down")) {
  
  # Get up/down-reg
  if (direction == "up") { 
    
    splineProbes <- splineDF[splineDF$logFC > 0,] 
    
    # ebyProbes <- probes.eby.bmicont.df[probes.eby.bmicont.df$logFC > 0,] 
    # trtProbes <- probes.trt.bmicont.df[probes.trt.bmicont.df$logFC > 0,] 
    # 
    # # CpG-level
    # print(paste0("RUNNING FOR :- ",direction))
    # up_goMethCpG_eby <- RunMethylEnrichment(ebyProbes, sig_cpgs=sig_cpgs, array.type="EPIC", cpg_df=probes.eby.bmicont.df[,c("Name","CONTRAST")],
    #                                            groupIsBackGroundSubset=TRUE, groupCol="CONTRAST",
    #                                         pval_col="adj.P.Val", dir=gocpg_dir, CpGCol="Name", extra=paste0(".",direction,"DiffCpGs"))
    # up_goMethCpG_trt <- RunMethylEnrichment(trtProbes, sig_cpgs=sig_cpgs, array.type="EPIC", cpg_df=probes.trt.bmicont.df[,c("Name","CONTRAST")],
    #                                            groupIsBackGroundSubset=TRUE, groupCol="CONTRAST",
    #                                         pval_col="adj.P.Val", dir=gocpg_dir, CpGCol="Name", extra=paste0(".",direction,"DiffCpGs_treatLog2FC0.1"))
    
    up_goMethCpG_trt <- RunMethylEnrichment(splineProbes, sig_cpgs=sig_cpgs, array.type="EPIC", cpg_df=splineDF[,c("Name","CONTRAST")],
                                               groupIsBackGroundSubset=TRUE, groupCol="CONTRAST",
                                            pval_col="bonferroni", dir=gocpg_dir, CpGCol="Name", extra=paste0(".",direction,"splineDiffCpGs_treatLog2FC0.1"))
     
  # ELse if down
  } else if (direction == "down") { 
    
     splineProbes <- splineDF[splineDF$logFC < 0,] 
    
    # ebyProbes <- probes.eby.bmicont.df[probes.eby.bmicont.df$logFC < 0,] 
    # trtProbes <- probes.trt.bmicont.df[probes.trt.bmicont.df$logFC < 0,] 
    # 
    # # CpG-level
    # print(paste0("RUNNING FOR :- ",direction))
    # down_goMethCpG_eby <- RunMethylEnrichment(ebyProbes, sig_cpgs=sig_cpgs, array.type="EPIC", cpg_df=probes.eby.bmicont.df[,c("Name","CONTRAST")],
    #                                            groupIsBackGroundSubset=TRUE, groupCol="CONTRAST",
    #                                           pval_col="adj.P.Val", dir=gocpg_dir, CpGCol="Name", extra=paste0(".",direction,"DiffCpGs"))
    # down_goMethCpG_trt <- RunMethylEnrichment(trtProbes, sig_cpgs=sig_cpgs, array.type="EPIC", cpg_df=probes.trt.bmicont.df[,c("Name","CONTRAST")],
    #                                            groupIsBackGroundSubset=TRUE, groupCol="CONTRAST",
    #                                           pval_col="adj.P.Val", dir=gocpg_dir, CpGCol="Name", extra=paste0(".",direction,"DiffCpGs_treatLog2FC0.1"))
    
         down_goMethCpG_trt <- RunMethylEnrichment(splineProbes, sig_cpgs=sig_cpgs, array.type="EPIC", cpg_df=splineDF[,c("Name","CONTRAST")],
                                               groupIsBackGroundSubset=TRUE, groupCol="CONTRAST",
                                            pval_col="bonferroni", dir=gocpg_dir, CpGCol="Name", extra=paste0(".",direction,"splineDiffCpGs_treatLog2FC0.1"))
    
    
  # Else if all comparisons
  } else if (direction == "all") {
    
    ebyProbes <- probes.eby.bmicont.df
    trtProbes <- probes.trt.bmicont.df

    # CpG-level
    print(paste0("RUNNING FOR :- ",direction))
    all_goMethCpG_eby <- RunMethylEnrichment(ebyProbes, sig_cpgs=sig_cpgs, array.type="EPIC", cpg_df=probes.eby.bmicont.df[,c("Name","CONTRAST")],
                                               groupIsBackGroundSubset=TRUE, groupCol="CONTRAST",
                                             pval_col="adj.P.Val", dir=gocpg_dir, CpGCol="Name", extra=paste0(".",direction,"DiffCpGs"))
    all_goMethCpG_trt <- RunMethylEnrichment(trtProbes, sig_cpgs=sig_cpgs, array.type="EPIC", cpg_df=probes.trt.bmicont.df[,c("Name","CONTRAST")],
                                               groupIsBackGroundSubset=TRUE, groupCol="CONTRAST",
                                             pval_col="adj.P.Val", dir=gocpg_dir, CpGCol="Name", extra=paste0(".",direction,"splineDiffCpGs_treatLog2FC0.1"))
    
  }
  
}

```

#### DMR

```{r}

# Split by up- and down-regulation
for (direction in c("all","up","down")) {
  
  # Get up/down-reg
  if (direction == "up") { 
    
    stageDMR <- stageTall_df[stageTall_df$logFC > 0,] 
    sexDMR <- sexTall_df[sexTall_df$logFC > 0,] 
    
    # CpG-level
    print(paste0("RUNNING FOR :- ",direction))
    up_goRegionDMRStage <- RunMethylEnrichment(stageDMR[stageDMR$p.adjust < 0.05,], cpg_df=probes.trt.bmicont.df[,c("Name","CONTRAST")], 
                                               groupIsBackGroundSubset=TRUE, groupCol="CONTRAST",
                                               sig_cpgs=c("Z_STAGE_WEEKS"=1), cpg_level=FALSE, array.type="EPIC",
                                               pval_col="p.adjust", dir=godmr_dir, CpGCol="Name", extra=paste0(".",direction,"DMRStage"))
    up_goRegionDMRSex   <- RunMethylEnrichment(sexDMR[sexDMR$p.adjust < 0.05,], cpg_df=probes.trt.bmicont.df[,c("Name","CONTRAST")], 
                                               groupIsBackGroundSubset=TRUE, groupCol="CONTRAST",
                                               sig_cpgs=c("sexmale - sexfemale"=1), cpg_level=FALSE, array.type="EPIC",
                                               pval_col="p.adjust", dir=godmr_dir, CpGCol="Name", extra=paste0(".",direction,"DMRSex"))
     
  # ELse if down
  } else if (direction == "down") { 
    
    stageDMR <- stageTall_df[stageTall_df$logFC < 0,] 
    sexDMR <- sexTall_df[sexTall_df$logFC < 0,] 
    
    # CpG-level
    print(paste0("RUNNING FOR :- ",direction))
    down_goRegionDMRStage <- RunMethylEnrichment(stageDMR[stageDMR$p.adjust < 0.05,], cpg_df=probes.trt.bmicont.df[,c("Name","CONTRAST")], 
                                                 groupIsBackGroundSubset=TRUE, groupCol="CONTRAST",
                                                 sig_cpgs=c("Z_STAGE_WEEKS"=1), cpg_level=FALSE, array.type="EPIC",
                                                 pval_col="p.adjust", dir=godmr_dir, CpGCol="Name", extra=paste0(".",direction,"DMRStage"))
    down_goRegionDMRSex   <- RunMethylEnrichment(sexDMR[sexDMR$p.adjust < 0.05,], cpg_df=probes.trt.bmicont.df[,c("Name","CONTRAST")], 
                                                 groupIsBackGroundSubset=TRUE, groupCol="CONTRAST",
                                                 sig_cpgs=c("sexmale - sexfemale"=1), cpg_level=FALSE, array.type="EPIC",
                                                 pval_col="p.adjust", dir=godmr_dir, CpGCol="Name", extra=paste0(".",direction,"DMRSex"))
    
  # Else if all comparisons
  } else if (direction == "all") {
    
    stageDMR <- stageTall_df
    sexDMR <- sexTall_df
    
    # CpG-level
    print(paste0("RUNNING FOR :- ",direction))
    all_goRegionDMRStage <- RunMethylEnrichment(stageDMR[stageDMR$p.adjust < 0.05,], cpg_df=probes.trt.bmicont.df[,c("Name","CONTRAST")], 
                                                groupIsBackGroundSubset=TRUE, groupCol="CONTRAST",
                                                sig_cpgs=c("Z_STAGE_WEEKS"=1), cpg_level=FALSE, array.type="EPIC",
                                                pval_col="p.adjust", dir=godmr_dir, CpGCol="Name", extra=paste0(".",direction,"DMRStage"))
    all_goRegionDMRSex   <- RunMethylEnrichment(sexDMR[sexDMR$p.adjust < 0.05,], cpg_df=probes.trt.bmicont.df[,c("Name","CONTRAST")], 
                                                groupIsBackGroundSubset=TRUE, groupCol="CONTRAST",
                                                sig_cpgs=c("sexmale - sexfemale"=1), cpg_level=FALSE, array.type="EPIC",
                                                pval_col="p.adjust", dir=godmr_dir, CpGCol="Name", extra=paste0(".",direction,"DMRSex"))
    
  }
  
}

```

### Plot

#### CpGs

```{r eval=FALSE}

PlotTopEnrichTerms <- function(df, sortCol, topN=10, outDir="./", extraDetails="_go", ...) {
  
  # If any sig terms
  if (nrow(df) > 0) {

    # Order
    df <- df[order(df[[sortCol]], decreasing=ifelse(sortCol == "FDR", FALSE, TRUE)),]
    
    # Subset, accounting for when less than topN sig terms
    if (nrow(df) >= topN) { df <- df[1:topN,] }
      
    # Plot
    plt <- ggplot(df, aes(x=reorder(TERM,-log10(FDR)), y=-log10(FDR), fill=FOLD_ENRICHMENT, size=log2(N))) + 
                  geom_point(stat="identity", shape=21) + 
                  scale_fill_continuous(high="red",low="blue") +
                  xlab("") +
                  theme_bw(base_size=16) + theme(axis.text.x=element_text(angle=90, hjust=1, vjust=0.5),
                                                 legend.position="top",
                                                 axis.title.y=element_text(size=24),legend.key.size = unit(1.5, "cm"))
    
    # Save
    ggsave(paste0(outDir, "dotPlot_top",topN,"Terms_by",sortCol,extraDetails,"Results.pdf"), plt, units="px", device="pdf", dpi=300, ...)
    ggsave(paste0(outDir, "dotPlot_top",topN,"Terms_by",sortCol,extraDetails,"Results.png"), plt, units="px", device="png", dpi=300, ...)
    
    # View plot
    return(plt)
    
  }
}

# Plot
## Stage
### All
PlotTopEnrichTerms(df=all_goMethCpG_trt[all_goMethCpG_trt$FDR < 0.05 & all_goMethCpG_trt$CONTRAST == "Z_STAGE_WEEKS" & all_goMethCpG_trt$ONTOLOGY == "BP",], 
                   sortCol="FDR", topN=10, outDir=gocpg_dir, extraDetails="_allStageCpGs_goBP", width=4000, height=3250)
PlotTopEnrichTerms(df=all_goMethCpG_trt[all_goMethCpG_trt$FDR < 0.05 & all_goMethCpG_trt$CONTRAST == "Z_STAGE_WEEKS" & all_goMethCpG_trt$ONTOLOGY == "MF",], 
                   sortCol="FDR", topN=10, outDir=gocpg_dir, extraDetails="_allStageCpGs_goMF", width=4000, height=3250)
PlotTopEnrichTerms(df=all_goMethCpG_trt[all_goMethCpG_trt$FDR < 0.05 & all_goMethCpG_trt$CONTRAST == "Z_STAGE_WEEKS" & all_goMethCpG_trt$ONTOLOGY == "CC",], 
                   sortCol="FDR", topN=10, outDir=gocpg_dir, extraDetails="_allStageCpGs_goCC", width=4000, height=3250)

### Up
PlotTopEnrichTerms(df=up_goMethCpG_trt[up_goMethCpG_trt$FDR < 0.05 & up_goMethCpG_trt$CONTRAST == "Z_STAGE_WEEKS" & up_goMethCpG_trt$ONTOLOGY == "BP",], 
                   sortCol="FDR", topN=10, outDir=gocpg_dir, extraDetails="_upStageCpGs_goBP", width=4000, height=3250)
PlotTopEnrichTerms(df=up_goMethCpG_trt[up_goMethCpG_trt$FDR < 0.05 & up_goMethCpG_trt$CONTRAST == "Z_STAGE_WEEKS" & up_goMethCpG_trt$ONTOLOGY == "MF",], 
                   sortCol="FDR", topN=10, outDir=gocpg_dir, extraDetails="_upStageCpGs_goMF", width=4000, height=3250)
PlotTopEnrichTerms(df=up_goMethCpG_trt[up_goMethCpG_trt$FDR < 0.05 & up_goMethCpG_trt$CONTRAST == "Z_STAGE_WEEKS" & up_goMethCpG_trt$ONTOLOGY == "CC",], 
                   sortCol="FDR", topN=10, outDir=gocpg_dir, extraDetails="_upStageCpGs_goCC", width=4000, height=3250)

### Down
PlotTopEnrichTerms(df=down_goMethCpG_trt[down_goMethCpG_trt$FDR < 0.05 & down_goMethCpG_trt$CONTRAST == "Z_STAGE_WEEKS" & down_goMethCpG_trt$ONTOLOGY == "BP",], 
                   sortCol="FDR", topN=10, outDir=gocpg_dir, extraDetails="_downStageCpGs_goBP", width=4000, height=3250)
PlotTopEnrichTerms(df=down_goMethCpG_trt[down_goMethCpG_trt$FDR < 0.05 & down_goMethCpG_trt$CONTRAST == "Z_STAGE_WEEKS" & down_goMethCpG_trt$ONTOLOGY == "MF",], 
                   sortCol="FDR", topN=10, outDir=gocpg_dir, extraDetails="_downStageCpGs_goMF", width=4000, height=3250)
PlotTopEnrichTerms(df=down_goMethCpG_trt[down_goMethCpG_trt$FDR < 0.05 & down_goMethCpG_trt$CONTRAST == "Z_STAGE_WEEKS" & down_goMethCpG_trt$ONTOLOGY == "CC",], 
                   sortCol="FDR", topN=10, outDir=gocpg_dir, extraDetails="_downStageCpGs_goCC", width=4000, height=3250)

## Sex
### All
PlotTopEnrichTerms(df=all_goMethCpG_eby[all_goMethCpG_eby$FDR < 0.05 & all_goMethCpG_eby$CONTRAST == "sexmale - sexfemale" & all_goMethCpG_eby$ONTOLOGY == "BP",], 
                   sortCol="FDR", topN=10, outDir=gocpg_dir, extraDetails="_allSexCpGs_goBP", width=4000, height=3250)
PlotTopEnrichTerms(df=all_goMethCpG_eby[all_goMethCpG_eby$FDR < 0.05 & all_goMethCpG_eby$CONTRAST == "sexmale - sexfemale" & all_goMethCpG_eby$ONTOLOGY == "MF",], 
                   sortCol="FDR", topN=10, outDir=gocpg_dir, extraDetails="_allSexCpGs_goMF", width=4000, height=3250)
PlotTopEnrichTerms(df=all_goMethCpG_eby[all_goMethCpG_eby$FDR < 0.05 & all_goMethCpG_eby$CONTRAST == "sexmale - sexfemale" & all_goMethCpG_eby$ONTOLOGY == "CC",], 
                   sortCol="FDR", topN=10, outDir=gocpg_dir, extraDetails="_allSexCpGs_goCC", width=4000, height=3250)

### Up
PlotTopEnrichTerms(df=up_goMethCpG_eby[up_goMethCpG_eby$FDR < 0.05 & up_goMethCpG_eby$CONTRAST == "sexmale - sexfemale" & up_goMethCpG_eby$ONTOLOGY == "BP",], 
                   sortCol="FDR", topN=10, outDir=gocpg_dir, extraDetails="_upSexCpGs_goBP", width=4000, height=3250)
PlotTopEnrichTerms(df=up_goMethCpG_eby[up_goMethCpG_eby$FDR < 0.05 & up_goMethCpG_eby$CONTRAST == "sexmale - sexfemale" & up_goMethCpG_eby$ONTOLOGY == "MF",], 
                   sortCol="FDR", topN=10, outDir=gocpg_dir, extraDetails="_upSexCpGs_goMF", width=4000, height=3250)
PlotTopEnrichTerms(df=up_goMethCpG_eby[up_goMethCpG_eby$FDR < 0.05 & up_goMethCpG_eby$CONTRAST == "sexmale - sexfemale" & up_goMethCpG_eby$ONTOLOGY == "CC",], 
                   sortCol="FDR", topN=10, outDir=gocpg_dir, extraDetails="_upSexCpGs_goCC", width=4000, height=3250)

### Down
PlotTopEnrichTerms(df=down_goMethCpG_eby[down_goMethCpG_eby$FDR < 0.05 & down_goMethCpG_eby$CONTRAST == "sexmale - sexfemale" & down_goMethCpG_eby$ONTOLOGY == "BP",], 
                   sortCol="FDR", topN=10, outDir=gocpg_dir, extraDetails="_downSexCpGs_goBP", width=4000, height=3250)
PlotTopEnrichTerms(df=down_goMethCpG_eby[down_goMethCpG_eby$FDR < 0.05 & down_goMethCpG_eby$CONTRAST == "sexmale - sexfemale" & down_goMethCpG_eby$ONTOLOGY == "MF",], 
                   sortCol="FDR", topN=10, outDir=gocpg_dir, extraDetails="_downSexCpGs_goMF", width=4000, height=3250)
PlotTopEnrichTerms(df=down_goMethCpG_eby[down_goMethCpG_eby$FDR < 0.05 & down_goMethCpG_eby$CONTRAST == "sexmale - sexfemale" & down_goMethCpG_eby$ONTOLOGY == "CC",], 
                   sortCol="FDR", topN=10, outDir=gocpg_dir, extraDetails="_downSexCpGs_goCC", width=4000, height=3250)

```

#### DMRs

```{r}

# Plot
## Stage
### All
PlotTopEnrichTerms(df=all_goRegionDMRStage[all_goRegionDMRStage$p.adjust < 0.05 & all_goRegionDMRStage$GROUP == "Z_STAGE_WEEKS" & all_goRegionDMRStage$ONTOLOGY == "BP",], 
                   sortCol="p.adjust", topN=10, outDir=godmr_dir, extraDetails="_allStageDMRs_goBP", width=4000, height=3250)
PlotTopEnrichTerms(df=all_goRegionDMRStage[all_goRegionDMRStage$p.adjust < 0.05 & all_goRegionDMRStage$GROUP == "Z_STAGE_WEEKS" & all_goRegionDMRStage$ONTOLOGY == "MF",], 
                   sortCol="p.adjust", topN=10, outDir=godmr_dir, extraDetails="_allStageDMRs_goMF", width=4000, height=3250)
PlotTopEnrichTerms(df=all_goRegionDMRStage[all_goRegionDMRStage$p.adjust < 0.05 & all_goRegionDMRStage$GROUP == "Z_STAGE_WEEKS" & all_goRegionDMRStage$ONTOLOGY == "CC",], 
                   sortCol="p.adjust", topN=10, outDir=godmr_dir, extraDetails="_allStageDMRs_goCC", width=4000, height=3250)

### Up
PlotTopEnrichTerms(df=up_goRegionDMRStage[up_goRegionDMRStage$p.adjust < 0.05 & up_goRegionDMRStage$GROUP == "Z_STAGE_WEEKS" & up_goRegionDMRStage$ONTOLOGY == "BP",], 
                   sortCol="p.adjust", topN=10, outDir=godmr_dir, extraDetails="_upStageDMRs_goBP", width=4000, height=3250)
PlotTopEnrichTerms(df=up_goRegionDMRStage[up_goRegionDMRStage$p.adjust < 0.05 & up_goRegionDMRStage$GROUP == "Z_STAGE_WEEKS" & up_goRegionDMRStage$ONTOLOGY == "MF",], 
                   sortCol="p.adjust", topN=10, outDir=godmr_dir, extraDetails="_upStageDMRs_goMF", width=4000, height=3250)
PlotTopEnrichTerms(df=up_goRegionDMRStage[up_goRegionDMRStage$p.adjust < 0.05 & up_goRegionDMRStage$GROUP == "Z_STAGE_WEEKS" & up_goRegionDMRStage$ONTOLOGY == "CC",], 
                   sortCol="p.adjust", topN=10, outDir=godmr_dir, extraDetails="_upStageDMRs_goCC", width=4000, height=3250)

### Down
PlotTopEnrichTerms(df=down_goRegionDMRStage[down_goRegionDMRStage$p.adjust < 0.05 & down_goRegionDMRStage$GROUP == "Z_STAGE_WEEKS" & down_goRegionDMRStage$ONTOLOGY == "BP",], 
                   sortCol="p.adjust", topN=10, outDir=godmr_dir, extraDetails="_downStageDMRs_goBP", width=4000, height=3250)
PlotTopEnrichTerms(df=down_goRegionDMRStage[down_goRegionDMRStage$p.adjust < 0.05 & down_goRegionDMRStage$GROUP == "Z_STAGE_WEEKS" & down_goRegionDMRStage$ONTOLOGY == "MF",], 
                   sortCol="p.adjust", topN=10, outDir=godmr_dir, extraDetails="_downStageDMRs_goMF", width=4000, height=3250)
PlotTopEnrichTerms(df=down_goRegionDMRStage[down_goRegionDMRStage$p.adjust < 0.05 & down_goRegionDMRStage$GROUP == "Z_STAGE_WEEKS" & down_goRegionDMRStage$ONTOLOGY == "CC",], 
                   sortCol="p.adjust", topN=10, outDir=godmr_dir, extraDetails="_downStageDMRs_goCC", width=4000, height=3250)

## Sex
### All
PlotTopEnrichTerms(df=all_goRegionDMRSex[all_goRegionDMRSex$p.adjust < 0.05 & all_goRegionDMRSex$GROUP == "sexmale - sexfemale" & all_goRegionDMRSex$ONTOLOGY == "BP",], 
                   sortCol="p.adjust", topN=10, outDir=godmr_dir, extraDetails="_allSexDMRs_goBP", width=4000, height=3250)
PlotTopEnrichTerms(df=all_goRegionDMRSex[all_goRegionDMRSex$p.adjust < 0.05 & all_goRegionDMRSex$GROUP == "sexmale - sexfemale" & all_goRegionDMRSex$ONTOLOGY == "MF",], 
                   sortCol="p.adjust", topN=10, outDir=godmr_dir, extraDetails="_allSexDMRs_goMF", width=4000, height=3250)
PlotTopEnrichTerms(df=all_goRegionDMRSex[all_goRegionDMRSex$p.adjust < 0.05 & all_goRegionDMRSex$GROUP == "sexmale - sexfemale" & all_goRegionDMRSex$ONTOLOGY == "CC",], 
                   sortCol="p.adjust", topN=10, outDir=godmr_dir, extraDetails="_allSexDMRs_goCC", width=4000, height=3250)

### Up
PlotTopEnrichTerms(df=up_goRegionDMRSex[up_goRegionDMRSex$p.adjust < 0.05 & up_goRegionDMRSex$GROUP == "sexmale - sexfemale" & up_goRegionDMRSex$ONTOLOGY == "BP",], 
                   sortCol="p.adjust", topN=10, outDir=godmr_dir, extraDetails="_upSexDMRs_goBP", width=4000, height=3250)
PlotTopEnrichTerms(df=up_goRegionDMRSex[up_goRegionDMRSex$p.adjust < 0.05 & up_goRegionDMRSex$GROUP == "sexmale - sexfemale" & up_goRegionDMRSex$ONTOLOGY == "MF",], 
                   sortCol="p.adjust", topN=10, outDir=godmr_dir, extraDetails="_upSexDMRs_goMF", width=4000, height=3250)
PlotTopEnrichTerms(df=up_goRegionDMRSex[up_goRegionDMRSex$p.adjust < 0.05 & up_goRegionDMRSex$GROUP == "sexmale - sexfemale" & up_goRegionDMRSex$ONTOLOGY == "CC",], 
                   sortCol="p.adjust", topN=10, outDir=godmr_dir, extraDetails="_upSexDMRs_goCC", width=4000, height=3250)

### Down
PlotTopEnrichTerms(df=down_goRegionDMRSex[down_goRegionDMRSex$p.adjust < 0.05 & down_goRegionDMRSex$GROUP == "sexmale - sexfemale" & down_goRegionDMRSex$ONTOLOGY == "BP",], 
                   sortCol="p.adjust", topN=10, outDir=godmr_dir, extraDetails="_downSexDMRs_goBP", width=4000, height=3250)
PlotTopEnrichTerms(df=down_goRegionDMRSex[down_goRegionDMRSex$p.adjust < 0.05 & down_goRegionDMRSex$GROUP == "sexmale - sexfemale" & down_goRegionDMRSex$ONTOLOGY == "MF",], 
                   sortCol="p.adjust", topN=10, outDir=godmr_dir, extraDetails="_downSexDMRs_goMF", width=4000, height=3250)
PlotTopEnrichTerms(df=down_goRegionDMRSex[down_goRegionDMRSex$p.adjust < 0.05 & down_goRegionDMRSex$GROUP == "sexmale - sexfemale" & down_goRegionDMRSex$ONTOLOGY == "CC",], 
                   sortCol="p.adjust", topN=10, outDir=godmr_dir, extraDetails="_downSexDMRs_goCC", width=4000, height=3250)

```

## sessionInfo

```{r}
writeLines(capture.output(sessionInfo()), paste0(log_dir,"methylation_sessionInfo.txt"))
```
